feat: Add optional output file name flag
This commit adds an optional `-output` flag to the `html2pdf` tool.
This flag allows the user to specify the name of the output PDF file.

If the `-output` flag is not provided, the tool will continue to
generate the file name from the title of the web page, preserving the
existing behavior.